requirements.txt
environs==9.5.0
requests==2.28.1
python-telegram-bot==13.14
